finished part 4

+++ part4/exercises/blog/backend/controllers/blogs.js
@@ -2,8 +2,19 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
+const jwt = require('jsonwebtoken')
+
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
+
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('userId', { name: 1, username: 1 });
   response.json(blogs);
 })
 
@@ -20,7 +31,14 @@ blogsRouter.get('/:id', async (request, response, next) => {
 blogsRouter.post('/',  async (request, response, next) => {
   const body = request.body;
 
-  const user = await User.findById(body.userId)
+  const token = getTokenFrom(request);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(body.userId);
 
   const blog = new Blog({
     title: body.title,
@@ -38,6 +56,15 @@ blogsRouter.post('/',  async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = getTokenFrom(request);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  const userId = decodedToken.id;
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.userId.toString() !== userId.toString()) {
+    return response.status(401).json({ error: 'only owners can delete blogs' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 })
